llmr/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-reasoner"
version = "0.1.0"
description = "LLM-based generative backend for NL → PartialDesignator → FullAction resolution"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "langchain-openai>=0.1",
    "langchain-ollama>=0.1",
    "langgraph>=0.1",
    "python-dotenv",
    "typing_extensions",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 99
target-version = ["py310", "py311", "py312"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["../test/llmr_test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
